Refractor panel mesh
